Remove OAuthIntrospectionResponse
The official TypeScript implementation contains no
evidence that oauth-introspection-response.ts has
been used anywhere. For this reason, the Swift
implementation of the file and its applicable
conversion documentation will be removed.

If, in the future, the official TypeScript implementation
decides to use anything in the file, the Swift
implementation and applicable conversion
documentation will come back.